backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.6.0
python-multipart==0.0.9
python-dotenv==1.0.0
